+++ qgreenland/tasks/layers/__init__.py
@@ -3,6 +3,7 @@ from qgreenland.tasks.layers.arctic_vegetation import ArcticVegetation
 from qgreenland.tasks.layers.background_image import BackgroundImage
 from qgreenland.tasks.layers.bedmachine import BedMachineDataset
 from qgreenland.tasks.layers.glacier_terminus import GlacierTerminus
+from qgreenland.tasks.layers.ice_thickness_change import IceThicknessChange
 from qgreenland.tasks.layers.rarred_shapefile import RarredShapefile
 from qgreenland.tasks.layers.velocity_mosaic import VelocityMosaic
 from qgreenland.tasks.layers.zipped_shapefile import ZippedShapefile
@@ -13,6 +14,7 @@ INGEST_TASKS = {
     'background_image': BackgroundImage,
     'bedmachine': BedMachineDataset,
     'glacier_terminus': GlacierTerminus,
+    'ice_thickness_change': IceThicknessChange,
     'rarred_shapefile': RarredShapefile,
     'velocity_mosaic': VelocityMosaic,
     'zipped_shapefile': ZippedShapefile